requirements.txt
streamlit==1.35.0
openai==1.33.0
pandas==2.2.2
plotly==5.22.0
python-dotenv==1.0.1
databricks-sql-connector==2.9.1
requests==2.31.0
